Restructure native file paths
Make file paths more closely match the non-native files, to make auto-complete file names slightly easier.
